+++ packages/forge-config/tests/index.spec.js
@@ -1,23 +1,21 @@
 const path = require('path');
-const { parseConfig } = require('../../');
+const { parse } = require('../lib');
 
-describe('#parseConfig', () => {
+describe('#parse', () => {
   test('should be a function', () => {
-    expect(typeof parseConfig).toEqual('function');
+    expect(typeof parse).toEqual('function');
   });
 
   test('should throw error if invalid file', () => {
     try {
-      parseConfig('./tmp/file');
+      parse('./tmp/file');
     } catch (err) {
       expect(err).toBeTruthy();
     }
   });
 
   test('should parse config as expected', () => {
-    const config = parseConfig(
-      path.resolve(__dirname, '../../../../examples/kvstore/config/forge.toml')
-    );
+    const config = parse(path.resolve(__dirname, '../../../examples/kvstore/config/forge.toml'));
     expect(config.app).toBeTruthy();
     expect(config.forge.decimal).toBeTruthy();
     expect(config.forge.web).toBeTruthy();